Don't raise a Puppet::ParseError in a report
Instructors were getting the following warning when the `irc.yaml` file
was inadvertently written to the wrong place.

```
Error: Could not send report: Error 400 on SERVER: Could not autoload puppet/reports/irc: Class Irc is already defined in Puppet::Reports
```

This appears to correct it. My best guess, having not crawled completely
through the code, is that instantiating a `Puppet::ParseError` somehow
causes more to be instantiated than expected.