Calculate defaults on more platforms
This assumes that the master the class is enforced on has the same
configuration as the master it's being compiled on--a sate bet, imho.
Just use the built in $settings hash to get the configured paths.

It also uses some pretty icky logic to determine the default gem
provider to use. I wish there was a better way to do this right now.